refactored code for abstraction with second player in TicTacToe (Issue #10)
