+++ src/edu/ucsb/cs56/projects/games/rock_paper_scissors/FirstPlayer.java
@@ -50,25 +50,28 @@ public class FirstPlayer extends JFrame {
         public void actionPerformed(ActionEvent e) {
             String selected = group.getSelection().getActionCommand();
             
+	    new SecondPlayer(selected);
+	    dispose();
+	    /*
             //make Squirtle image for first player
             if (selected == "Squirtle"){
-                new NoSquirtle();
+		//new NoSquirtle();
 		dispose();
             }
             
             //make bulbasaur image for first player
             if (selected == "Bulbasaur"){
-                new NoBulbasaur();
+                //new NoBulbasaur();
                 dispose();
             }
             
             //make Charmander image for first player
             if (selected == "Charmander"){
-                new NoCharmander();
+                //new NoCharmander();
                 dispose();
             }
+	    */
 
-      
        }
     }
